\name{mwa_data}
\docType{data}
\alias{mwa_data}
\title{data to illustrate the functionality of MWA}
\description{
  This artificial data set illustrates how MWA can be used to identitfy causal effects. Treatment, control, and dependent events are referenced in time and space. Increased levels of dependent events following treatments can be visually and numerically analyzed with MWA.  
}
\usage{data(mwa_data)}
\format{A data.frame containing observations.}
\source{Monte Carlo Simulations. See supplementary information of Schutte and Donnay (2014).}
\references{
Schutte, S., Donnay, K. (2014). ``Matched wake analysis: Finding causal relationships in spatiotemporal event data.'' \emph{Political Geography} 41:1-10.
}
%\keyword{datasets}
